Fix Next.js 15 async searchParams and client/server boundary issues
- Make searchParams async and await it properly in page.tsx (Next.js 15 requirement)
- Move parseLogEntries to shared utils.ts to fix client/server module boundary error
- Update tests to import from utils.ts
- Avoid code duplication by using single parseLogEntries function for both client and server

This fixes the following errors:
1. "searchParams should be awaited before using its properties"
2. "Attempted to call parseLogEntries() from the server but parseLogEntries is on the client"
